# web/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_preview_scan_report_import_route_api_work_projects__project_id__scan_report_imports_preview_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_preview_scan_report_import_route_api_work_projects__project_id__scan_report_imports_preview_post",
        "type": "object"
      },
...
      "CommitScanReportImportRequest": {
        "additionalProperties": false,
        "properties": {
          "create_blackboard_fact": {
            "default": true,
            "title": "Create Blackboard Fact",
            "type": "boolean"
          },
          "create_relationships": {
            "default": true,
            "title": "Create Relationships",
            "type": "boolean"
          },
          "selected_asset_keys": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "maxItems": 110000,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Asset Keys"
          }
        },
        "title": "CommitScanReportImportRequest",
        "type": "object"
      },
      "CommitScanReportImportResponse": {
        "properties": {
          "blackboard_node_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blackboard Node Id"
          },
          "created_assets": {
            "items": {
              "type": "string"
            },
            "title": "Created Assets",
            "type": "array"
          },
          "created_relationships": {
            "items": {
              "type": "string"
            },
            "title": "Created Relationships",
            "type": "array"
          },
          "import_id": {
            "title": "Import Id",
            "type": "string"
          },
          "skipped_assets": {
            "items": {
              "type": "string"
            },
            "title": "Skipped Assets",
            "type": "array"
          },
          "skipped_relationships": {
            "items": {
              "type": "string"
            },
            "title": "Skipped Relationships",
            "type": "array"
          },
          "updated_assets": {
            "items": {
              "type": "string"
            },
            "title": "Updated Assets",
            "type": "array"
          }
        },
        "required": [
          "import_id",
          "created_assets",
          "updated_assets",
          "skipped_assets",
          "created_relationships",
          "skipped_relationships"
        ],
        "title": "CommitScanReportImportResponse",
        "type": "object"
      },
      "CommonResponse": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse",
        "type": "object"
      },
...
      "CommonResponse_Any_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "title": "Data"
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[Any]",
        "type": "object"
      },
...
      "CommonResponse_CommitScanReportImportResponse_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CommitScanReportImportResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[CommitScanReportImportResponse]",
        "type": "object"
      },
...
      "CommonResponse_ScanReportImportPreviewResponse_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScanReportImportPreviewResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[ScanReportImportPreviewResponse]",
        "type": "object"
      },
...
      "ScanReportAssetCandidate": {
        "additionalProperties": false,
        "properties": {
          "extra": {
            "$ref": "#/components/schemas/WorkProjectAssetExtraSchema"
          },
          "host": {
            "title": "Host",
            "type": "string"
          },
          "identifier": {
            "title": "Identifier",
            "type": "string"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "origin": {
            "$ref": "#/components/schemas/WorkProjectAssetOrigin",
            "default": "discovered"
          },
          "path": {
            "default": "",
            "title": "Path",
            "type": "string"
          },
          "port": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Port"
          },
          "type": {
            "$ref": "#/components/schemas/WorkProjectAssetType"
          }
        },
        "required": [
          "key",
          "type",
          "identifier",
          "host"
        ],
        "title": "ScanReportAssetCandidate",
        "type": "object"
      },
      "ScanReportImportCounts": {
        "properties": {
          "assets": {
            "default": 0,
            "title": "Assets",
            "type": "integer"
          },
          "domains": {
            "default": 0,
            "title": "Domains",
            "type": "integer"
          },
          "hosts": {
            "default": 0,
            "title": "Hosts",
            "type": "integer"
          },
          "networks": {
            "default": 0,
            "title": "Networks",
            "type": "integer"
          },
          "relationships": {
            "default": 0,
            "title": "Relationships",
            "type": "integer"
          },
          "services": {
            "default": 0,
            "title": "Services",
            "type": "integer"
          }
        },
        "title": "ScanReportImportCounts",
        "type": "object"
      },
      "ScanReportImportFormat": {
        "enum": [
          "nmap_xml"
        ],
        "title": "ScanReportImportFormat",
        "type": "string"
      },
      "ScanReportImportInfo": {
        "properties": {
          "committed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Committed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "default": "",
            "title": "Error",
            "type": "string"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "format": {
            "$ref": "#/components/schemas/ScanReportImportFormat"
          },
          "import_id": {
            "title": "Import Id",
            "type": "string"
          },
          "media_type": {
            "title": "Media Type",
            "type": "string"
          },
          "project_id": {
            "title": "Project Id",
            "type": "integer"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/ScanReportImportStatus"
          },
          "summary": {
            "additionalProperties": true,
            "title": "Summary",
            "type": "object"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "uploader_user_id": {
            "title": "Uploader User Id",
            "type": "integer"
          }
        },
        "required": [
          "import_id",
          "project_id",
          "uploader_user_id",
          "filename",
          "media_type",
          "sha256",
          "format",
          "status",
          "size_bytes",
          "created_at",
          "updated_at"
        ],
        "title": "ScanReportImportInfo",
        "type": "object"
      },
      "ScanReportImportPreviewResponse": {
        "properties": {
          "assets": {
            "items": {
              "$ref": "#/components/schemas/ScanReportAssetCandidate"
            },
            "title": "Assets",
            "type": "array"
          },
          "counts": {
            "$ref": "#/components/schemas/ScanReportImportCounts"
          },
          "duplicate_keys": {
            "items": {
              "type": "string"
            },
            "title": "Duplicate Keys",
            "type": "array"
          },
          "import_info": {
            "$ref": "#/components/schemas/ScanReportImportInfo"
          },
          "relationships": {
            "items": {
              "$ref": "#/components/schemas/ScanReportRelationshipCandidate"
            },
            "title": "Relationships",
            "type": "array"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "import_info",
          "assets",
          "relationships",
          "warnings",
          "duplicate_keys",
          "counts"
        ],
        "title": "ScanReportImportPreviewResponse",
        "type": "object"
      },
      "ScanReportImportStatus": {
        "enum": [
          "previewed",
          "committed",
          "failed"
        ],
        "title": "ScanReportImportStatus",
        "type": "string"
      },
      "ScanReportRelationshipCandidate": {
        "additionalProperties": false,
        "properties": {
          "key": {
            "title": "Key",
            "type": "string"
          },
          "source_asset_key": {
            "title": "Source Asset Key",
            "type": "string"
          },
          "target_asset_key": {
            "title": "Target Asset Key",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/WorkProjectGraphEdgeType"
          }
        },
        "required": [
          "key",
          "source_asset_key",
          "target_asset_key",
          "type"
        ],
        "title": "ScanReportRelationshipCandidate",
        "type": "object"
      },
...
      "WorkProjectAssetExtraSchema": {
        "additionalProperties": false,
        "description": "Recon metadata attached to an asset. Kept small on purpose.",
        "properties": {
          "banner": {
            "default": "",
            "maxLength": 512,
            "title": "Banner",
            "type": "string"
          },
          "protocol": {
            "default": "tcp",
            "maxLength": 16,
            "minLength": 1,
            "title": "Protocol",
            "type": "string"
          },
          "service_name": {
            "default": "",
            "maxLength": 128,
            "title": "Service Name",
            "type": "string"
          }
        },
        "title": "WorkProjectAssetExtraSchema",
        "type": "object"
      },
      "WorkProjectAssetOrigin": {
        "enum": [
          "scope",
          "discovered"
        ],
        "title": "WorkProjectAssetOrigin",
        "type": "string"
      },
...
      "WorkProjectAssetType": {
        "enum": [
          "service",
          "domain",
          "network",
          "binary"
        ],
        "title": "WorkProjectAssetType",
        "type": "string"
      },
...
      "WorkProjectGraphEdgeType": {
        "enum": [
          "related",
          "resolves_to",
          "hosts",
          "connects_to",
          "trusts",
          "exploits",
          "pivots_to",
          "leads_to"
        ],
        "title": "WorkProjectGraphEdgeType",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "AccessTokenAuth": {
        "in": "header",
        "name": "X-XM-Access-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "XuanMu RedTeam Agent - 开源红队多智能体协作平台，用于授权安全评估、代码审计、渗透测试与安全研究。",
    "version": "0.2.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/work-projects/{project_id}/scan-report-imports/preview": {
      "post": {
        "operationId": "preview_scan_report_import_route_api_work_projects__project_id__scan_report_imports_preview_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_preview_scan_report_import_route_api_work_projects__project_id__scan_report_imports_preview_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_ScanReportImportPreviewResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Work project not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "Preview Scan Report Import Route",
        "tags": [
          "work-project-scan-report-imports"
        ]
      }
    },
    "/api/work-projects/{project_id}/scan-report-imports/{import_id}/commit": {
      "post": {
        "operationId": "commit_scan_report_import_route_api_work_projects__project_id__scan_report_imports__import_id__commit_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "import_id",
            "required": true,
            "schema": {
              "title": "Import Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommitScanReportImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_CommitScanReportImportResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Work project or scan report import not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "Commit Scan Report Import Route",
        "tags": [
          "work-project-scan-report-imports"
        ]
      }
    }
  }
}